@@ -4,39 +4,38 @@
   "private": true,
   "dependencies": {
     "@bentley/icons-generic": "^1.0.34",
-    "@bentley/react-scripts": "^5.0.4",
-    "@itwin/appui-abstract": "^3.7.0",
-    "@itwin/appui-layout-react": "^3.7.0",
-    "@itwin/appui-react": "^3.7.0",
-    "@itwin/browser-authorization": "^0.5.1",
-    "@itwin/components-react": "^3.7.0",
-    "@itwin/core-bentley": "^3.7.0",
-    "@itwin/core-common": "^3.7.0",
-    "@itwin/core-frontend": "^3.7.0",
-    "@itwin/core-geometry": "^3.7.0",
-    "@itwin/core-i18n": "^3.7.0",
-    "@itwin/core-markup": "^3.7.0",
-    "@itwin/core-orbitgt": "^3.7.0",
-    "@itwin/core-quantity": "^3.7.0",
-    "@itwin/core-react": "^3.7.0",
-    "@itwin/core-telemetry": "^3.7.0",
-    "@itwin/imodel-components-react": "^3.7.0",
-    "@itwin/imodels-access-frontend": "^1.0.2",
-    "@itwin/itwinui-react": "^1.48.2",
-    "@itwin/presentation-common": "^3.7.0",
-    "@itwin/presentation-frontend": "^3.7.0",
-    "@itwin/web-viewer-react": "^3.1.4",
-    "@itwin/webgl-compatibility": "^3.7.0",
-    "@types/history": "^4.7.11",
-    "@types/node": "^14.18.42",
-    "@types/react": "^17.0.55",
-    "@types/react-dom": "^17.0.19",
-    "history": "^4.10.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-redux": "^7.2.9",
-    "redux": "^4.2.1",
-    "typescript": "~4.4.0"
+    "@bentley/react-scripts": "^5.0.7",
+    "@itwin/appui-abstract": "^4.5.2",
+    "@itwin/appui-react": "^4.12.0",
+    "@itwin/browser-authorization": "^1.1.0",
+    "@itwin/components-react": "^4.12.0",
+    "@itwin/core-bentley": "^4.5.2",
+    "@itwin/core-common": "^4.5.2",
+    "@itwin/core-frontend": "^4.5.2",
+    "@itwin/core-geometry": "^4.5.2",
+    "@itwin/core-i18n": "^4.5.2",
+    "@itwin/core-markup": "^4.5.2",
+    "@itwin/core-orbitgt": "^4.5.2",
+    "@itwin/core-quantity": "^4.5.2",
+    "@itwin/core-react": "^4.12.0",
+    "@itwin/core-telemetry": "^4.5.2",
+    "@itwin/imodel-components-react": "^4.12.0",
+    "@itwin/imodels-access-frontend": "^4.0.0",
+    "@itwin/itwinui-react": "^3.10.0",
+    "@itwin/presentation-common": "^4.5.2",
+    "@itwin/presentation-components": "^5.2.2",
+    "@itwin/presentation-frontend": "^4.5.2",
+    "@itwin/tree-widget-react": "^2.2.0",
+    "@itwin/web-viewer-react": "^4.2.4",
+    "@itwin/webgl-compatibility": "^4.5.2",
+    "@types/node": "^20.12.8",
+    "@types/react": "^18.3.1",
+    "@types/react-dom": "^18.3.0",
+    "history": "^5.3.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.23.0",
+    "typescript": "^5.4.5"
   },
   "scripts": {
     "start": "react-scripts --max_old_space_size=4096 start",
