@@ -3,14 +3,11 @@
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
 
-import "./index.scss";
-
-import { BrowserAuthorizationCallbackHandler } from "@itwin/browser-authorization";
+import "@itwin/itwinui-react/styles.css";
 import React from "react";
-import ReactDOM from "react-dom";
-
-import App from "./App";
-import * as serviceWorker from "./serviceWorker";
+import { createRoot } from "react-dom/client";
+import { App } from "./components/App";
+import "./index.scss";
 
 if (!process.env.IMJS_AUTH_CLIENT_CLIENT_ID) {
   throw new Error(
@@ -28,21 +25,9 @@ if (!process.env.IMJS_AUTH_CLIENT_REDIRECT_URI) {
   );
 }
 
-const redirectUrl = new URL(process.env.IMJS_AUTH_CLIENT_REDIRECT_URI);
-if (redirectUrl.pathname === window.location.pathname) {
-  BrowserAuthorizationCallbackHandler.handleSigninCallback(
-    redirectUrl.toString()
-  ).catch(console.error);
-} else {
-  ReactDOM.render(
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>,
-    document.getElementById("root")
-  );
-}
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+const root = createRoot(document.getElementById("root")!);
+root.render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>
+);
